move assert-invariants from migrate -> cmd
